remove duplications and test versions
