+++ Common/Core/CMakeLists.txt
@@ -23,6 +23,7 @@ o2physics_target_root_dictionary(AnalysisCore
                         PID/DetectorResponse.h
                         PID/PIDTOF.h
                         PID/TOFReso.h
+                        PID/PIDTPC.h
                         PID/TPCPIDResponse.h
               LINKDEF AnalysisCoreLinkDef.h)
